requirements.txt
pytest==8.4.1
pytest-asyncio==1.1.0
requests==2.32.5
pytest-mock==3.14.1
httpx==0.28.1
fastapi==0.116.1
pytest-cov==6.2.1
pytest-benchmark==5.1.0
aiohttp==3.12.15
tortoise==0.1.1
testcontainers==4.12.0
tortoise-orm==0.25.1
asyncpg==0.30.0
python-jose==3.5.0



